+++ backend/src/neuroagent/app/dependencies.py
@@ -299,6 +299,7 @@ def get_s3_client(
 
 
 def get_context_variables(
+    request: Request,
     settings: Annotated[Settings, Depends(get_settings)],
     starting_agent: Annotated[Agent, Depends(get_starting_agent)],
     httpx_client: Annotated[AsyncClient, Depends(get_httpx_client)],
@@ -310,6 +311,7 @@ def get_context_variables(
     """Get the context variables to feed the tool's metadata."""
     return {
         "bluenaas_url": settings.tools.bluenaas.url,
+        "brainregion_embeddings": request.app.state.br_embeddings,
         "brainregion_hierarchy_storage_key": settings.storage.brain_region_hierarchy_key,
         "bucket_name": settings.storage.bucket_name,
         "celltypes_hierarchy_storage_key": settings.storage.cell_type_hierarchy_key,