+++ backend/tests/app/routers/test_threads.py
@@ -62,7 +62,7 @@ def test_generate_thread_title(
 
     with app_client as app_client:
         threads = app_client.get("/threads/").json()
-        assert not threads
+        assert not threads["results"]
 
         create_output_1 = app_client.post(
             "/threads",
@@ -101,7 +101,7 @@ def test_get_threads(patch_required_env, httpx_mock, app_client, db_connection):
 
     with app_client as app_client:
         threads = app_client.get("/threads").json()
-        assert not threads
+        assert not threads["results"]
         create_output_1 = app_client.post(
             "/threads",
             json={"virtual_lab_id": "test_vlab", "project_id": "test_project"},
@@ -113,24 +113,24 @@ def test_get_threads(patch_required_env, httpx_mock, app_client, db_connection):
         create_output_3 = app_client.post("/threads").json()
         threads = app_client.get("/threads").json()
 
-        assert len(threads) == 1
-        assert threads[0] == create_output_3
+        assert len(threads["results"]) == 1
+        assert threads["results"][0] == create_output_3
 
         threads = app_client.get(
             "/threads",
             params={"virtual_lab_id": "test_vlab", "project_id": "test_project"},
         ).json()
 
-        assert len(threads) == 1
-        assert threads[0] == create_output_1
+        assert len(threads["results"]) == 1
+        assert threads["results"][0] == create_output_1
 
         threads = app_client.get(
             "/threads",
             params={"virtual_lab_id": "test_vlab2", "project_id": "test_project2"},
         ).json()
 
-        assert len(threads) == 1
-        assert threads[0] == create_output_2
+        assert len(threads["results"]) == 1
+        assert threads["results"][0] == create_output_2
 
         threads = app_client.get(
             "/threads",
@@ -138,6 +138,48 @@ def test_get_threads(patch_required_env, httpx_mock, app_client, db_connection):
         )
 
         assert threads.status_code == 401
+
+
+@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
+def test_get_threads_paginated(
+    patch_required_env, httpx_mock, app_client, db_connection
+):
+    mock_keycloak_user_identification(httpx_mock)
+    test_settings = Settings(
+        db={"prefix": db_connection}, keycloak={"issuer": "https://great_issuer.com"}
+    )
+    app.dependency_overrides[get_settings] = lambda: test_settings
+
+    with app_client as app_client:
+        threads = app_client.get("/threads").json()
+        assert set(threads.keys()) == {
+            "next_cursor",
+            "has_more",
+            "page_size",
+            "results",
+        }
+
+        create_output_1 = app_client.post(
+            "/threads",
+        ).json()
+        create_output_2 = app_client.post(
+            "/threads",
+        ).json()
+        create_output_3 = app_client.post("/threads").json()
+        threads = app_client.get("/threads", params={"page_size": 2}).json()
+
+        assert threads["page_size"] == 2
+        assert threads["next_cursor"] == create_output_2["update_date"]
+        assert threads["has_more"]
+        assert len(threads["results"]) == 2
+
+        assert threads["results"][0]["thread_id"] == create_output_3["thread_id"]
+        assert threads["results"][1]["thread_id"] == create_output_2["thread_id"]
+
+        page_2 = app_client.get(
+            "/threads", params={"page_size": 2, "cursor": threads["next_cursor"]}
+        ).json()
+        assert page_2["results"][0] == create_output_1
 
 
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
@@ -152,7 +194,7 @@ def test_get_threads_query_param(
 
     with app_client as app_client:
         threads = app_client.get("/threads").json()
-        assert not threads
+        assert not threads["results"]
         # Create some threads
         create_output_1 = app_client.post("/threads").json()
         create_output_2 = app_client.post("/threads").json()
@@ -167,20 +209,22 @@ def test_get_threads_query_param(
             create_output_3["thread_id"],
         ]
         threads = app_client.get("/threads", params={"sort": "-creation_date"}).json()
-        assert [thread["thread_id"] for thread in threads] == list(reversed(thread_ids))
+        assert [thread["thread_id"] for thread in threads["results"]] == list(
+            reversed(thread_ids)
+        )
 
         threads = app_client.get("/threads", params={"sort": "creation_date"}).json()
-        assert [thread["thread_id"] for thread in threads] == thread_ids
+        assert [thread["thread_id"] for thread in threads["results"]] == thread_ids
 
         threads = app_client.get("/threads", params={"sort": "-update_date"}).json()
-        assert [thread["thread_id"] for thread in threads] == [
+        assert [thread["thread_id"] for thread in threads["results"]] == [
             thread_ids[1],
             thread_ids[2],
             thread_ids[0],
         ]
 
         threads = app_client.get("/threads", params={"sort": "update_date"}).json()
-        assert [thread["thread_id"] for thread in threads] == [
+        assert [thread["thread_id"] for thread in threads["results"]] == [
             thread_ids[0],
             thread_ids[2],
             thread_ids[1],
@@ -197,7 +241,7 @@ def test_update_thread_title(patch_required_env, httpx_mock, app_client, db_conn
 
     with app_client as app_client:
         threads = app_client.get("/threads").json()
-        assert not threads
+        assert not threads["results"]
 
         # Check when wrong thread id
         wrong_response = app_client.patch(
@@ -238,7 +282,7 @@ def test_delete_thread(
 
     with app_client as app_client:
         threads = app_client.get("/threads").json()
-        assert not threads
+        assert not threads["results"]
 
         # Check when wrong thread id
         wrong_response = app_client.delete("/threads/wrong_id")
@@ -255,14 +299,14 @@ def test_delete_thread(
             "/threads",
             params={"virtual_lab_id": "test_vlab", "project_id": "test_project"},
         ).json()
-        assert len(threads) == 1
-        assert threads[0]["thread_id"] == thread_id
+        assert len(threads["results"]) == 1
+        assert threads["results"][0]["thread_id"] == thread_id
 
         delete_response = app_client.delete(f"/threads/{thread_id}").json()
         assert delete_response["Acknowledged"] == "true"
 
         threads = app_client.get("/threads").json()
-        assert not threads
+        assert not threads["results"]
 
         assert fake_delete_from_storage.call_count == 1
 
@@ -287,7 +331,9 @@ async def test_get_thread_messages(
 
     with app_client as app_client:
         # Get the messages of the thread
-        messages = app_client.get(f"/threads/{thread.thread_id}/messages").json()
+        messages = app_client.get(
+            f"/threads/{thread.thread_id}/messages", params={"sort": "creation_date"}
+        ).json()["results"]
 
     assert messages[0]["entity"] == "user"
     assert messages[0]["msg_content"] == {"content": "This is my query."}
@@ -340,7 +386,7 @@ async def test_get_thread_messages_sort_and_filter(
             f"/threads/{thread.thread_id}/messages",
             params={"sort": "-creation_date", "entity": ["USER", "TOOL"]},
         )
-        messages = response.json()
+        messages = response.json()["results"]
 
     # Expecting only the messages that have the entities "user" and "tool".
     # From the populate_db fixture these are:
@@ -363,7 +409,7 @@ async def test_get_thread_messages_sort_and_filter(
             f"/threads/{thread.thread_id}/messages",
             params={"sort": "creation_date", "entity": ["AI_TOOL", "AI_MESSAGE"]},
         )
-        messages = response.json()
+        messages = response.json()["results"]
 
     # Expecting only the messages that have the entities "ai_tool" and "ai_message".
     # According to populate_db these are:
@@ -380,3 +426,94 @@ async def test_get_thread_messages_sort_and_filter(
     assert messages[0]["msg_content"] == {"content": ""}
     assert messages[1]["entity"] == "ai_message"
     assert messages[1]["msg_content"] == {"content": "sample response content."}
+
+
+@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
+@pytest.mark.asyncio
+async def test_get_thread_messages_paginated(
+    patch_required_env,
+    httpx_mock,
+    app_client,
+    db_connection,
+    populate_db,
+):
+    mock_keycloak_user_identification(httpx_mock)
+    test_settings = Settings(
+        db={"prefix": db_connection}, keycloak={"issuer": "https://great_issuer.com"}
+    )
+    app.dependency_overrides[get_settings] = lambda: test_settings
+
+    db_items, _ = populate_db
+    thread = db_items["thread"]
+
+    with app_client as app_client:
+        # Get the messages of the thread
+        messages = app_client.get(
+            f"/threads/{thread.thread_id}/messages", params={"page_size": 3}
+        ).json()
+        page_2 = app_client.get(
+            f"/threads/{thread.thread_id}/messages",
+            params={"page_size": 3, "cursor": messages["next_cursor"]},
+        ).json()
+
+    assert set(messages.keys()) == {"next_cursor", "has_more", "page_size", "results"}
+
+    assert messages["page_size"] == 3
+    assert messages["next_cursor"] == messages["results"][-1]["creation_date"]
+    assert messages["has_more"]
+    assert len(messages["results"]) == 3
+
+    messages_results = messages["results"]
+
+    assert messages_results[2]["entity"] == "ai_tool"
+    assert messages_results[2]["msg_content"] == {"content": ""}
+    assert messages_results[2]["message_id"]
+    assert messages_results[2]["creation_date"]
+
+    assert messages_results[1]["entity"] == "tool"
+    assert messages_results[1]["msg_content"] == {"content": "It's sunny today."}
+    assert messages_results[1]["message_id"]
+    assert messages_results[1]["creation_date"]
+
+    assert messages_results[0]["entity"] == "ai_message"
+    assert messages_results[0]["msg_content"] == {"content": "sample response content."}
+    assert messages_results[0]["message_id"]
+    assert messages_results[0]["creation_date"]
+
+    assert messages_results[0]["creation_date"] > messages_results[1]["creation_date"]
+    assert messages_results[1]["creation_date"] > messages_results[2]["creation_date"]
+
+    assert len(page_2["results"]) == 1
+
+
+@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
+@pytest.mark.asyncio
+async def test_get_thread_messages_empty_paginated(
+    patch_required_env,
+    httpx_mock,
+    app_client,
+    db_connection,
+):
+    mock_keycloak_user_identification(httpx_mock)
+    test_settings = Settings(
+        db={"prefix": db_connection}, keycloak={"issuer": "https://great_issuer.com"}
+    )
+    app.dependency_overrides[get_settings] = lambda: test_settings
+
+    with app_client as app_client:
+        # Create a thread
+        create_output = app_client.post(
+            "/threads",
+        ).json()
+        thread = create_output["thread_id"]
+        # Get the messages of the thread
+        messages = app_client.get(
+            f"/threads/{thread}/messages", params={"page_size": 3}
+        ).json()
+
+    assert set(messages.keys()) == {"next_cursor", "has_more", "page_size", "results"}
+
+    assert messages["page_size"] == 3
+    assert messages["next_cursor"] is None
+    assert not messages["has_more"]
+    assert len(messages["results"]) == 0