+++ CHANGELOG.md
@@ -8,12 +8,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
-- Get morpho tool and resolve entities now uses entitycore.
 - Bearer token is propagated through the `httpx_client`.
+- Use `.model_dump(exclude_defaults=True)` for auto-generated classes dumping.
+- Get morpho tool and resolve entities now uses entitycore.
+- Remove usage of user journey in suggestion after first message.
+- Tools based on API calls adapted to new autogen tool.
 
 ### Added
 - Autogeneration of input schemas from APIs.
 - Morphometrics tool using obi-one.
+- Custom brain region resolving.
+
+### Fixed
+- Handle breaking change in entitycore.
 
 ## [0.5.4]
 
@@ -63,6 +70,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Filter and sort get_thread and get_thread/messages.
 - Plots for the SCS simulations.
 - New semantic route.
+- Pagination for threads and messages.
 
 ### Fixed
 - Literature search points to right url.